WEBHOOKS READY TO DEPLOY

+++ events/messageCreate.js
@@ -28,35 +28,33 @@ module.exports = {
       }
     })
 
-    try {
-      let webhooks = await message.channel.fetchWebhooks()
-      let webhook = webhooks.find(wh => wh.owner.id == client.user.id)
+    let webhooks = await message.channel.fetchWebhooks()
+    let webhook = webhooks.find(wh => wh.owner.id == client.user.id)
 
-      let args = message.content.slice(prefix.length).trim().split(/ +/g);
-      let content = args.slice(0).join(" ").replace(prefix);
-      message.delete()
+    if (typeof prefix == 'undefined') {
+      return console.log('nope!')
+    }
+    let args = message.content.slice(prefix.length).trim().split(/ +/g);
+    let content = args.slice(0).join(" ").replace(prefix);
+    message.delete()
 
-      if (typeof webhook == 'undefined') {
-        message.channel.createWebhook(`${message.channel.name}`, { avatar: client.user.displayAvatarURL() }).then(wb => {
-          wb.send({
-            content: content,
-            username: webhookName,
-            avatarURL: imgURL,
-          });
-        })
-      } else {
-        await webhook.send({
+    if (typeof webhook == 'undefined') {
+      message.channel.createWebhook(`${message.channel.name}`, { avatar: client.user.displayAvatarURL() }).then(wb => {
+        wb.send({
           content: content,
           username: webhookName,
           avatarURL: imgURL,
         });
-      }
-
-    } catch (err) {
-      console.log(err)
+      })
+    } else {
+      await webhook.send({
+        content: content,
+        username: webhookName,
+        avatarURL: imgURL,
+      });
     }
 
-    let args = message.content.slice(config.prefix.length).trim().split(/ +/g);
+    args = message.content.slice(config.prefix.length).trim().split(/ +/g);
 
     if (message.content.slice(0, config.prefix.length) !== config.prefix) return;
     const cmdName = args.shift().toLowerCase();
